requirements.txt
absl-py==2.1.0
altair==5.3.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dateutils==0.6.12
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozendict==2.4.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
html5lib==1.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
multitasking==0.0.11
mysqlclient==2.2.4
namex==0.0.8
nsepython==2.8
nsetools==1.0.11
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
peewee==3.17.3
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
streamlit==1.34.0
tenacity==8.3.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
yfinance==0.2.38
